refactor: 토스페이먼츠 쿠폰 적용 추가
- User: 비즈니스 로직 Type Long으로 변경
- TossPaymentsService: 쿠폰 로직 추가하여 수정

+++ src/main/java/com/example/demo/user/entity/User.java
@@ -50,7 +50,7 @@ public class User {
     }
 
     /* 비즈니스 로직*/
-    public void addPoint(int amount) {
+    public void addPoint(Long amount) {
         balance += amount;
     }
 }
